refactor: extract browser automation commands
